Add files via upload
This project based on  the movie industry, is Netflix the new entertainment king? 
Those were the first questions that lead me to create a datasetfocused on movie revenue and analyze it over the last decades. But, why stop there? There are more factors that intervene in thiskind of thing, like actors, genres, user ratings and more. And now, anyone with experience (you) can ask specific questions about the
movie industry, and get answers.
The Correlation in python in which i use jupyter notebook, libraries of python like pandas, seaborn, matplotlib, numpy.
With the help of these libraries i used to clean the data like removing the duplicates values, collect the data in descending or ascending order that helpful to analyze the in better way etc.